sdk/requirements.txt
# Core
Django==3.2.5
Jinja2==3.0.1
underscore.py==0.1.6
validators==0.18.2
PyYAML==5.4.1
pika==1.2.0
simplejson==3.17.3

# Rest framework
djangorestframework==3.12.4
django-jsonfield==1.4.1
django-url-filter==0.3.15
requests==2.26.0
django-filter==2.4.0

psutil==5.6.3
retry==0.9.2

tabulate==0.8.9
